Math Library Updated
Math methods to operate with matrices implemented. E.g. Add two matrices, multiply two matrices, multiply a matrix for a scalar, get the identity of a specified dimension, get the inverse of a matrix, get the determinant of a matrix, get the transpose of a matrix, etc. And 4 dimension operations supported.